add useDrawer
