+++ packages/lab/src/drawer/index.ts
@@ -1,1 +1,2 @@
 export * from "./Drawer";
+export * from "./useDrawer";
